Some minor cleaning (#413)
* Some maintenance on the code.

* Removing warnings.

* Documenting the new 64-bit data structure.

* Adding actions.

+++ jmh/src/jmh/java/org/roaringbitmap/realdata/wrapper/RoaringBitmapWrapper.java
@@ -61,7 +61,7 @@ final class RoaringBitmapWrapper implements Bitmap {
 
   @Override
   public Bitmap flip(int rangeStart, int rangeEnd) {
-    return new RoaringBitmapWrapper(RoaringBitmap.flip(bitmap, rangeStart, rangeEnd));
+    return new RoaringBitmapWrapper(RoaringBitmap.flip(bitmap, (long)rangeStart, (long)rangeEnd));
   }
