+++ jmh/src/jmh/java/org/roaringbitmap/runcontainer/RunArrayOrBenchmark.java
@@ -73,7 +73,7 @@ public class RunArrayOrBenchmark {
 
         for (int z = 0; z < 50; ++z) {
           int end = start + rand.nextInt(10000);
-          rb.add(start, end);
+          rb.add((long)start, (long)end);
           start = end + rand.nextInt(1000);
         }
         cc.add(toConcise(rb.toArray()));
